% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_rr.R
\name{animate_rr}
\alias{animate_rr}
\title{\code{animate_rr} draws the path around the ring-road
using data \code{LF} from \code{rrtrials}.}
\usage{
animate_rr(LF)
}
\arguments{
\item{LF, }{a matrix}
}
\description{
\code{animate_rr} draws the path around the ring-road
using data \code{LF} from \code{rrtrials}.
}
