% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{\code{anova} uses the law of large numbers to establish confidence intervals traffic performance measures.}
\usage{
anova(zipper.output, sbs.output)
}
\arguments{
\item{zipper.output}{output from \code{zipper.simulate}, a data frame}

\item{sbs.output}{output from \code{sbs.simulate}, a data.frame}
}
\value{
The \code{anova} runs \code{brktrials4wrapper} \code{size} times.
}
\description{
\code{anova} uses the law of large numbers to establish confidence intervals traffic performance measures.
}
