\name{basic-package}
\alias{basic-package}
\alias{basic}
\docType{package}
\title{
\packageTitle{basic}
}
\description{
\packageDescription{basic}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{basic}
\packageIndices{basic}
}
\author{
\packageAuthor{basic}

Maintainer: \packageMaintainer{basic}
}
\keyword{ package }
