% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic}
\alias{basic}
\title{Basic interpreter}
\usage{
basic(
  basic_file = NULL,
  home_dir = system.file("scripts", package = "basic"),
  debug = rly::NullLogger$new()
)
}
\arguments{
\item{basic_file}{Name of basic file in the scripts folder (default NULL gives interpreter command prompt)}

\item{home_dir}{home directory for basic files}

\item{debug}{rly debugger}
}
\description{
Basic interpreter
}
\examples{
b <- basic('hello.bas')
b$run()
}
