% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASIC.function.R
\name{basic_axisQTL}
\alias{basic_axisQTL}
\title{Title}
\usage{
basic_axisQTL(beta, se, gamma_all, MDS_ALL, weight, nPC)
}
\arguments{
\item{beta}{This matrix contains eQTL effect sizes for both bulk and single-cell data. Column 1: Gene-SNP pair identifiers. Column 2: Bulk effect size. Columns 3+: Cell type-specific eQTL effect sizes.}

\item{se}{This matrix has the same dimensions as the effect size matrix and represents the standard errors associated with the eQTL effect sizes. Column 1: Gene-SNP pair identifiers (should match the effect size matrix). Column 2: Bulk effect size standard error. Columns 3+: Standard errors for the cell type-specific eQTL effect sizes.}

\item{gamma_all}{Output from meta_regression_fast}

\item{MDS_ALL}{Output from meta_regression_fast}

\item{weight}{Weights for each cell types}
}
\description{
Title
}
\examples{
basic_function(beta,se,gamma_all,MDS_ALL,weight,nPC)
}
