% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic_shell}
\alias{basic_shell}
\title{Interactive Basic interpreter}
\usage{
basic_shell(debug = rly::NullLogger$new())
}
\arguments{
\item{debug}{rly debugger}
}
\description{
Interactive Basic interpreter
}
\examples{
basic()
}
