% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_interpreter.R
\name{BasicInterpreter}
\alias{BasicInterpreter}
\title{BasicInterpreter object}
\description{
Class providing runtime support for running a basic program.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{functions}}{(\code{list()})\cr
Function table.}

\item{\code{vars}}{(\code{environment()})\cr
All variables.}

\item{\code{lists}}{(\code{list()})\cr
List variables.}

\item{\code{tables}}{(\code{matrix()})\cr
Tables.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BasicInterpreter-new}{\code{BasicInterpreter$new()}}
\item \href{#method-BasicInterpreter-collect_data}{\code{BasicInterpreter$collect_data()}}
\item \href{#method-BasicInterpreter-check_end}{\code{BasicInterpreter$check_end()}}
\item \href{#method-BasicInterpreter-check_loops}{\code{BasicInterpreter$check_loops()}}
\item \href{#method-BasicInterpreter-eval}{\code{BasicInterpreter$eval()}}
\item \href{#method-BasicInterpreter-releval}{\code{BasicInterpreter$releval()}}
\item \href{#method-BasicInterpreter-assign}{\code{BasicInterpreter$assign()}}
\item \href{#method-BasicInterpreter-goto}{\code{BasicInterpreter$goto()}}
\item \href{#method-BasicInterpreter-run}{\code{BasicInterpreter$run()}}
\item \href{#method-BasicInterpreter-expr_str}{\code{BasicInterpreter$expr_str()}}
\item \href{#method-BasicInterpreter-relexpr_str}{\code{BasicInterpreter$relexpr_str()}}
\item \href{#method-BasicInterpreter-var_str}{\code{BasicInterpreter$var_str()}}
\item \href{#method-BasicInterpreter-list}{\code{BasicInterpreter$list()}}
\item \href{#method-BasicInterpreter-erase}{\code{BasicInterpreter$erase()}}
\item \href{#method-BasicInterpreter-add_statements}{\code{BasicInterpreter$add_statements()}}
\item \href{#method-BasicInterpreter-del_line}{\code{BasicInterpreter$del_line()}}
\item \href{#method-BasicInterpreter-clone}{\code{BasicInterpreter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-new"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-new}{}}}
\subsection{Method \code{new()}}{
Initialize the interpreter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$new(prog)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prog}}{An environment containing (line, statement) mappings.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-collect_data"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-collect_data}{}}}
\subsection{Method \code{collect_data()}}{
Collect all data statements.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$collect_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-check_end"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-check_end}{}}}
\subsection{Method \code{check_end()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$check_end()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-check_loops"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-check_loops}{}}}
\subsection{Method \code{check_loops()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$check_loops()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-eval"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-eval}{}}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$eval(expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-releval"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-releval}{}}}
\subsection{Method \code{releval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$releval(expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-assign"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-assign}{}}}
\subsection{Method \code{assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$assign(target, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-goto"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-goto}{}}}
\subsection{Method \code{goto()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$goto(linenum)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-run"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-expr_str"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-expr_str}{}}}
\subsection{Method \code{expr_str()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$expr_str(expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-relexpr_str"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-relexpr_str}{}}}
\subsection{Method \code{relexpr_str()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$relexpr_str(expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-var_str"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-var_str}{}}}
\subsection{Method \code{var_str()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$var_str(var)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-list"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-list}{}}}
\subsection{Method \code{list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-erase"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-erase}{}}}
\subsection{Method \code{erase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$erase()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-add_statements"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-add_statements}{}}}
\subsection{Method \code{add_statements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$add_statements(prog)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-del_line"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-del_line}{}}}
\subsection{Method \code{del_line()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$del_line(lineno)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicInterpreter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BasicInterpreter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicInterpreter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
