\name{bb.mix3.plot}
\alias{bb.mix3.plot}
\title{
plot the results of mixture of three beta-binomials
}
\description{
plot the observed dentisy, overlaid with the densities of the data simulated
based on the of mixture of three beta-binomials 
}
\usage{
bb.mix3.plot(fit0)
}
\arguments{
  \item{fit0}{
  the output of function bb.mix3
}
}
\author{ Wei Sun weisun2013@gmail.com }
\keyword{ utilities }
