% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmfree.R
\name{bmfree}
\alias{bmfree}
\title{The \code{bmfree} function returns a \code{tux} matrix of time \code{t}, speed \code{u} and location \code{x} values.}
\usage{
bmfree(umn, usd, T, dt)
}
\arguments{
\item{umn}{mean speed (mph), a number}

\item{usd}{standard deviation of \code{umn}, a number}

\item{T}{upper time range, a number}

\item{dt}{time-step, a number}
}
\value{
The \code{bmfree} function returns information for a single vehicle
used to analyze the effects of traffic noise on car-following.
The function uses a stochastic traffic noise model, a brownian motion or Weiner \code{W} model.
}
\description{
The \code{bmfree} function returns a \code{tux} matrix of time \code{t}, speed \code{u} and location \code{x} values.
}
\examples{
bmfree(34, 3, 60, 5)
bmfree(2, 2, 60, 5)
}
