% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmfree3.R
\name{bmfree3}
\alias{bmfree3}
\title{The \code{bmfree3} function returns a \code{tux} matrix of time \code{t}, speed \code{u} (fps)
 and location \code{x} (feet).}
\usage{
bmfree3(ustart, uend, usd, tstart, tend, xstart, delt, type)
}
\arguments{
\item{ustart}{mean speed (mph), a number}

\item{uend}{mean speed (mph), a number}

\item{usd}{standard deviation of \code{usd} (mph), a number}

\item{tstart}{simulation start time, a number}

\item{tend}{simulation end time, a number}

\item{xstart}{location (feet), a number}

\item{delt}{time-step size, a number}

\item{type}{logical for plotting where TRUE creates a plot}
}
\description{
The \code{bmfree3} function returns a \code{tux} matrix of time \code{t}, speed \code{u} (fps)
 and location \code{x} (feet).
}
