% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakdownplot.R
\name{breakdownplot}
\alias{breakdownplot}
\title{The function \code{breakdownplot} plots the data from the \code{accelpass} data frame.}
\usage{
breakdownplot(df, title)
}
\arguments{
\item{df}{from \code{accelpass}, a matrix}

\item{title, }{a character string}
}
\description{
The function \code{breakdownplot} plots the data from the \code{accelpass} data frame.
}
