% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkcross0.R
\name{brkcross0}
\alias{brkcross0}
\title{\code{brkcross0} estimates of speed and location of vehicle crossing bottleneck location \code{x}}
\usage{
brkcross0(i, dftest)
}
\arguments{
\item{i}{index, a number}

\item{dftest, }{a dataframe}
}
\description{
\code{brkcross0} estimates of speed and location of vehicle crossing bottleneck location \code{x}
}
