% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brktrials.R
\name{brktrials}
\alias{brktrials}
\title{\code{brktrials} produces \code{t-x} trajectories for lead and following vehicles at a bottleneck}
\usage{
brktrials(tend, umn, usd, xstart, xfunnel, leff, lane, step)
}
\arguments{
\item{tend}{end time for simualted run, a number}

\item{umn}{start speed (mph) for vehicle in lane 1, a number}

\item{usd}{speed volatility for \code{umn}, a number}

\item{xstart}{start location for vehicle in lane 1 (feet), a vector}

\item{xfunnel}{upstream location where the lane drop starts (feet), a number}

\item{leff}{effective vehicle length(feet), a number}

\item{lane}{number, a vector}

\item{step}{size in seconds, a number}
}
\value{
\code{brktrials} returns  \code{t-x} tractories
for \code{nveh} vehicles at a bottleneck.
\code{k}, \code{xstart} and \code{lane} are input vectors of length \code{nveh}.
\code{brktrials} is a wrapper function of \code{accelpass}.
}
\description{
\code{brktrials} produces \code{t-x} trajectories for lead and following vehicles at a bottleneck
}
\examples{
brktrials(30, 41, 11, xstart, -500, 14, lane, 0.5)
}
