% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity.R
\name{capacity}
\alias{capacity}
\title{\code{capacity} is an algorithm for estimating the headway from \code{brk4trials} output}
\usage{
capacity(df1df2, tstart, tend, step, xfunnel)
}
\arguments{
\item{df1df2}{leading vehicle, a matrix}

\item{tstart}{start time, (seconds), a number}

\item{tend}{end time, (seconds), a number}

\item{step}{size in seconds, a number
xfunnel upstream location of bottleneck taper, a number}
}
\value{
\code{capacity} returns the mean and standard deviation of a capacity estimate.
}
\description{
\code{capacity} is an algorithm for estimating the headway from \code{brk4trials} output
}
\examples{
capacity(df1df2, tstart, tend, step, xfunnel)
}
