% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfamin.R
\name{cfanim}
\alias{cfanim}
\title{\code{cfanim} uses a Brownian motion model of speed to animate a driver's ability to maintain a speed \code{u}.}
\usage{
cfanim(u, usd, tup, dt)
}
\arguments{
\item{u}{mean speed (mph), a number}

\item{usd}{standard deviation of \code{umn}, a number}

\item{tup}{upper time range, a number}

\item{dt}{time-step, a number}
}
\value{
The \code{cfanim} function returns information for a single vehicle
used to analyze the effects of traffic noise on car-following.
The function uses a stochastic traffic noise model, a brownian motion or Weiner \code{W} model.
}
\description{
\code{cfanim} uses a Brownian motion model of speed to animate a driver's ability to maintain a speed \code{u}.
}
\examples{
cfanim(60, 3, 30, 0.25)
}
