% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chancemodel.R
\name{chancemodel}
\alias{chancemodel}
\title{\code{chancemodel} uses the law of large numbers to establish confidence intervals traffic performance measures.}
\usage{
chancemodel(outputruns, output)
}
\arguments{
\item{outputruns}{is a collection of \code{output} runs, a data frame.}

\item{output}{\code{simulate} output, a data frame}
}
\value{
The \code{chancemodel} calculates the mean3 and standard deviations from \code{simulate}.
}
\description{
\code{chancemodel} uses the law of large numbers to establish confidence intervals traffic performance measures.
}
