% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterbottleneck.R
\name{enterbottleneck}
\alias{enterbottleneck}
\title{\code{enterbottleneck} produces a data frame that lists the order vehicles enter the bottleneck at location \code{xfunnel}..}
\usage{
enterbottleneck(lane1,lane2, xfunnel, tend, step)
}
\arguments{
\item{lane1, }{a matrix}

\item{lane2, }{a matrix}

\item{xfunnel, }{a number}

\item{tend}{upper time range of simulation, a number}

\item{step}{time step, a number}
}
\value{
A list containing the table and number of vehicles \code{nveh}
}
\description{
\code{enterbottleneck} produces a data frame that lists the order vehicles enter the bottleneck at location \code{xfunnel}..
}
