% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialdesign2.R
\name{factorialdesign}
\alias{factorialdesign}
\title{\code{factorialdesign} .}
\usage{
factorialdesign(zip1, sbs1, zip2, sbs2, colnum, figtitle)
}
\arguments{
\item{zip1}{output from \code{zipper.simulate}, a data frame}

\item{sbs1}{output from \code{sbs.simulate}, a data.frame}

\item{zip2}{output from \code{zipper.simulate}, a data frame}

\item{sbs2}{output from \code{sbs.simulate}, a data.frame}

\item{colnum}{column number of interest, a number}

\item{figtitle, }{plot title, character string}
}
\value{
The \code{factorialdesign} runs \code{brktrials4wrapper} \code{size} times.
}
\description{
\code{factorialdesign} .
}
