% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findviolation.R
\name{findviolation}
\alias{findviolation}
\title{\code{findviolation} determines if a zone violates the safe headway rule.}
\usage{
findviolation(tstart, tend, tend.0, df1, df2, step, leff)
}
\arguments{
\item{tstart}{start time, a number}

\item{tend}{end time, a number}

\item{tend.0}{end time for over the long time range, a number}

\item{df1}{leading vehicle, a matrix}

\item{df2}{following vehicle, a matrix}

\item{step}{time step, a number}

\item{leff}{vehicle length, a number}
}
\value{
\code{findviolation} fills the cell of the \code{dfcrit} table.
}
\description{
\code{findviolation} determines if a zone violates the safe headway rule.
}
