% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixdf1df2.R
\name{fixdf1df2}
\alias{fixdf1df2}
\title{\code{fixdf1df2} returns a matrix with following vehicle information.}
\usage{
fixdf1df2(veh, df2.fix, df1df2)
}
\arguments{
\item{veh, }{a number}

\item{df2.fix}{output from \code{fixviolation}, matrix}

\item{df1df2, }{a matrix}
}
\value{
A matrix of speed, location data by time.
}
\description{
\code{fixdf1df2} returns a matrix with following vehicle information.
}
