% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixviolation.R
\name{fixviolation}
\alias{fixviolation}
\title{\code{fixviolation} produces of a safe headways for a following vehicle}
\usage{
fixviolation(veh, zone, df1df2, dfcrit, step, tend.0, leff, xfunnel)
}
\arguments{
\item{veh}{vehicle, a number}

\item{zone}{from \code{dfcrit} matrix for vehicle \code{veh}, a number}

\item{df1df2}{leading and following vehicle information, a matrix}

\item{dfcrit}{critical times, a matrix}

\item{step}{time step, a number}

\item{tend.0}{end time for over the long time range, a number}

\item{leff}{effective vehicle lenght, a number}

\item{xfunnel}{location where the lane drop is located, a number}
}
\value{
\code{fixviolation} is creates following vehicle trajectory that satisfies the safe headway rule.
}
\description{
\code{fixviolation} produces of a safe headways for a following vehicle
}
