% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixviolation2.R
\name{fixviolation2}
\alias{fixviolation2}
\title{\code{fixviolation2} produces of a safe headways for a following vehicle}
\usage{
fixviolation2(veh, zone, df1df2, dfcrit, delt, tend.0, leff, xfunnel, browse)
}
\arguments{
\item{veh}{vehicle, a number}

\item{zone}{from \code{dfcrit} matrix for vehicle \code{veh}, a number}

\item{df1df2}{leading and following vehicle information, a matrix}

\item{dfcrit}{critical time, a matrix}

\item{delt}{time step, a number}

\item{tend.0}{end time for over the long time range, a number}

\item{leff}{effective vehicle length in feet, a number}

\item{xfunnel}{location where the lane drop is located, a number}

\item{browse}{to inspect \code{fixviolation2} plot or FALSE otherwise}
}
\value{
\code{fixviolation2} is creates following vehicle trajectory that satisfies the safe headway rule.
}
\description{
\code{fixviolation2} produces of a safe headways for a following vehicle
}
