% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follower.R
\name{follower}
\alias{follower}
\title{\code{follower} returns a matrix of the following vehicle}
\usage{
follower(veh, df1df2)
}
\arguments{
\item{veh, }{a number}

\item{df1df2, }{a matrix}
}
\value{
A matrix of speed, location data by time.
}
\description{
\code{follower} returns a matrix of the following vehicle
}
