% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formqueue.R
\name{formqueue}
\alias{formqueue}
\title{The \code{formqueue} function is used to explain queuing upstream of a bottleneck.}
\usage{
formqueue(tstart, tend, umn, usd, xstart, xfunnel, leff)
}
\arguments{
\item{tstart}{start time, a number}

\item{tend}{end time, a number}

\item{umn}{start speed (mph) for vehicle in lane 1, a number}

\item{usd}{speed volatility for \code{umn}, a number}

\item{xstart}{start location for vehicle in lane 1 (feet), a number}

\item{xfunnel}{upstream location where the lane drop starts (feet), a number}

\item{leff}{effective vehicle lingth(feet), a number}
}
\value{
\code{formqueue} uses a deterministic model to illustrate an ``idealistic'' situation,
a so-called a ``formqueue merge.''
}
\description{
The \code{formqueue} function is used to explain queuing upstream of a bottleneck.
}
