% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.R
\name{gbm}
\alias{gbm}
\title{Estimate speed \code{u} at time \code{t} using a geometric Brownian motion model.}
\usage{
gbm(u0, theta, t)
}
\arguments{
\item{u0}{initial speed (fps), a number}

\item{theta, }{a number}

\item{t}{time (seconds), a number}
}
\description{
Estimate speed \code{u} at time \code{t} using a geometric Brownian motion model.
}
\examples{
gbm(92.4, -0.0154, 40)
gbm(0, 0.4, 4)
}
