\name{h1}
\alias{h1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Space headway \code{h1} at time \code{t1}, the time the following vehicle initiates deceleration.
}
\description{
\code{h1} is the minimum headway of based on the following vehicle sightline and \code{hsafe(u1)} where \code{u1} is the lead vehicle speed at \code{t} = \code{t1}.
}

\value{
\item{h1}{headway between lead and following vehicles, measured in feet.}
}
\author{
Paul J Ossenbruggen
}
