\name{h2}
\alias{h2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Space headway \code{h2} at time \code{t2}, the time the following vehicle driver sees the lead vehicle has decelerated.
}
\description{
\code{h2} is the headway \code{hsafe(u2)} where \code{u2} is the lead vehicle speed at \code{t} = \code{t2}.
}

\value{
\item{h2}{headway between lead and following vehicles, measured in feet.}
}
\author{
Paul J Ossenbruggen
}
