% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsafe.R
\name{hsafe}
\alias{hsafe}
\title{Safe headway \code{h} between lead and following vehicles.}
\usage{
hsafe(u, leff)
}
\arguments{
\item{u}{speed in units of fps, a number}

\item{leff}{effective vehicle length in feet, a number}
}
\value{
A safe headway is defined by a following vehicle speed equal to or less than its leading vehicle speed and
distance headway that is one passenger car length for each 10 mph.
}
\description{
Safe headway \code{h} between lead and following vehicles.
}
\examples{
hsafe(60, 14)
}
