% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I93speed.R
\docType{data}
\name{I93speed}
\alias{I93speed}
\title{Pre-breakdown and breakdown speed data from I-93 in Salem, New Hampshire in 2010.}
\format{A data frame with 8697 rows and 2 columns:
\describe{
\item{u}{traffic speed, miles per hour (mph)}
\item{t}{time index}
}}
\usage{
I93speed
}
\description{
A data frame of nortbound traffic speed observations that are aggregated over fifteen minute intervals.
The pre-breakdown speeds are shown as at times \code{t} = -1. The breakdown speeds are show at times \code{t} = 0,1,2,... .
}
\keyword{datasets}
