% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I93weekday.R
\docType{data}
\name{I93weekday}
\alias{I93weekday}
\title{Weekday speed, flow and density data from I-93 in Salem, New Hampshire in 2010.}
\format{A data frame with 19153 rows and 3 columns:
\describe{
\item{q}{traffic flow, in vehicles per hour (vph)}
\item{u}{traffic speed, miles per hour (mph)}
\item{k}{traffic density, vehicles per mile per lane (vpmpl)}
}}
\usage{
I93weekday
}
\description{
A dataset of nortbound traffic observations that are aggregated over fifteen minute intervals.
A complete record for a day will consist of 96 observations for a twenty-four hour day.
}
\keyword{datasets}
