% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealtrajectory.R
\name{idealtrajectory}
\alias{idealtrajectory}
\title{\code{idealtrajectory} creates a \code{t-x} trajectory for a deterministic model.}
\usage{
idealtrajectory(df1df2, nveh, u, k)
}
\arguments{
\item{df1df2}{leading vehicle, a matrix}

\item{nveh}{number of vehicles, a number}

\item{u}{speed in mph, a number}

\item{k}{density in vpm, a number}
}
\value{
\code{idealtrajectory} returns the mean and standard deviation of a idealtrajectory estimate.
}
\description{
\code{idealtrajectory} creates a \code{t-x} trajectory for a deterministic model.
}
