% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfactor.R
\docType{data}
\name{kfactor}
\alias{kfactor}
\title{\code{kfactor} used to adjust the density at time \code{t} = 0, a number}
\format{\code{kfactor}, for example, equals 4/3 to obtain density = 45 vpm at \code{t} = 0.
\describe{
\item{kfactor}{a number}
}}
\usage{
kfactor
}
\description{
\code{kfactor} used to adjust the density at time \code{t} = 0, a number
}
\keyword{datasets}
