\name{leff}
\alias{leff}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
\code{leff} is the effective length of a vehicle.
}
\description{
\code{leff} is the length of a vehicle.
}

\value{
\item{leff}{effective vehicle length, measured in feet.}
}
\author{
Paul J Ossenbruggen
}
