% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticmodel.R
\name{logisticmodel}
\alias{logisticmodel}
\title{The function \code{logisticmodel} estimates the parameters of logistic regression model
using a binomial formula and logit link function for data \code{QUKtdayX}.}
\usage{
logisticmodel(QUKtdayX)
}
\arguments{
\item{QUKtdayX, }{a matrix}
}
\description{
The function \code{logisticmodel} estimates the parameters of logistic regression model
using a binomial formula and logit link function for data \code{QUKtdayX}.
}
