% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedata.R
\name{mergedata}
\alias{mergedata}
\title{\code{mergedata} produces a data frame that lists the order vehicles enter the bottleneck at location \code{xfunnel}..}
\usage{
mergedata(lane1,lane2, tlen, dfcrit)
}
\arguments{
\item{lane1, }{a matrix}

\item{lane2, }{a matrix}

\item{tlen}{number of time steps, a number}

\item{dfcrit}{list of information about lanes and vehicles, a matrix}
}
\value{
\code{df1df2} is a matrix of both lane 1 and 2.
}
\description{
\code{mergedata} produces a data frame that lists the order vehicles enter the bottleneck at location \code{xfunnel}..
}
