% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASIC.function.R
\name{meta_regression_fast}
\alias{meta_regression_fast}
\title{Title}
\usage{
meta_regression_fast(beta, se, nPC)
}
\arguments{
\item{beta}{This matrix contains eQTL effect sizes for both bulk and single-cell data. Column 1: Gene-SNP pair identifiers. Column 2: Bulk effect size. Columns 3+: Cell type-specific eQTL effect sizes.}

\item{se}{This matrix has the same dimensions as the effect size matrix and represents the standard errors associated with the eQTL effect sizes. Column 1: Gene-SNP pair identifiers (should match the effect size matrix). Column 2: Bulk effect size standard error. Columns 3+: Standard errors for the cell type-specific eQTL effect sizes.}

\item{nPC}{Number fo cell types}
}
\description{
Title
}
\examples{
meta_regression_fast(beta,se,nPC)
}
