% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nopass.R
\name{nopass}
\alias{nopass}
\title{\code{nopass} produces \code{t}, \code{u} and \code{x} for lead following vehicles downstream of a bottleneck}
\usage{
nopass(nope, merge.df, step)
}
\arguments{
\item{veh}{vehicle, a number}

\item{nope}{is a data.frame of leading and following time, speed and location data, a dataframe}

\item{merge.df}{is a data.frame of leading and following time, speed and location data, a dataframe}

\item{step}{size in seconds, a number}
}
\value{
\code{brktrials} returns  a data frame for a vehicle at a bottleneck.
\code{brktrials} is a wrapper function of \code{accelpass}.
}
\description{
\code{nopass} produces \code{t}, \code{u} and \code{x} for lead following vehicles downstream of a bottleneck
}
