% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nveh1.R
\docType{data}
\name{nveh1}
\alias{nveh1}
\title{\code{nveh1} is the number of vehicles traveling in lane 1.}
\format{\code{nveh1} number of vehicles.
\describe{
\item{nveh1}{a number}
}}
\usage{
nveh1
}
\description{
\code{nveh1} is the number of vehicles traveling in lane 1.
}
\keyword{datasets}
