% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nveh2.R
\docType{data}
\name{nveh2}
\alias{nveh2}
\title{\code{nveh2} is the number of vehicles traveling in lane 2.}
\format{\code{nveh2} number of vehicles.
\describe{
\item{nveh2}{a number}
}}
\usage{
nveh2
}
\description{
\code{nveh2} is the number of vehicles traveling in lane 2.
}
\keyword{datasets}
