% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passplot.R
\name{passplot}
\alias{passplot}
\title{The function \code{passplot} plots the data from the \code{freeflowpass} or \code{decelmerge} data frames.}
\usage{
passplot(df, title)
}
\arguments{
\item{df}{from \code{freeflowpass}, a matrix}

\item{title, }{a character string}
}
\description{
The function \code{passplot} plots the data from the \code{freeflowpass} or \code{decelmerge} data frames.
}
