% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{\code{performance} looks at waiting time and performance measures.}
\usage{
performance(sbs.summary,zipper.summary,sbs.prop.free,zipper.prop.free,size)
}
\arguments{
\item{sbs.summary, }{a matrix}

\item{zipper.summary, }{a matrix}

\item{sbs.prop.free}{proportion free flow, a number}

\item{zipper.prop.free}{proportion free flow, a number}
}
\value{
The \code{performance} returns statistical summaries for two data sets.
}
\description{
\code{performance} looks at waiting time and performance measures.
}
