% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlane.R
\name{plotlane}
\alias{plotlane}
\title{Produces a time-distance \code{t-x} trajectory for the lead vehicle, \code{vehicle = 1}.}
\usage{
plotlane(lane)
}
\arguments{
\item{lane, }{a matrix}
}
\description{
Produces a time-distance \code{t-x} trajectory for the lead vehicle, \code{vehicle = 1}.
}
