% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlanes1and2.R
\name{plotlanes1and2}
\alias{plotlanes1and2}
\title{\code{plotlanes1and2} creates \code{t-x} trajectories for lanes 1 and 2.}
\usage{
plotlanes1and2(lane1, lane2, tend, delt, xfunnel)
}
\arguments{
\item{lane1}{from \code{brktrials3 <- lst[[1]]}, a matrix}

\item{lane2}{from \code{brktrials3 <- lst[[2]]}, a matrix}

\item{tend}{end time, (seconds), a number}

\item{delt}{size in seconds, a number}

\item{xfunnel, }{start location of the bottleneck, a number}
}
\description{
\code{plotlanes1and2} creates \code{t-x} trajectories for lanes 1 and 2.
}
