% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottux.R
\name{plottux}
\alias{plottux}
\title{\code{plottux} produces \code{t-u} and \code{t-x} trajectories for a lead and following vehicles on a ring road,}
\usage{
plottux(umn, usd, N, T, k0, leff, lambda)
}
\arguments{
\item{umn}{mean speed (mph), a number}

\item{usd}{standard deviation of \code{umn}, a number}

\item{N}{number of time-steps, a number}

\item{T}{upper time range in seconds, a number}

\item{k0}{traffic density (vpm), a number}

\item{leff}{effective vehicle length, a number}

\item{lambda}{calibration constant, a number}
}
\description{
\code{plottux} produces \code{t-u} and \code{t-x} trajectories for a lead and following vehicles on a ring road,
}
\examples{
plottux(41, 11, 120, 120, 55, 14, 1.25)
plottux(18.8, 3.8, 120, 120, 55, 14, 0.5)
}
