\name{print.bb.mix}
\alias{print.bb.mix}
\title{
  print a bb.mix object
}
\description{
  print parameter estimate and a snapshot of input data for a bb.mix object.
}
\usage{
print.bb.mix(x, ...)
}
\arguments{
  \item{x}{ a bb.mimx object, which is the output of function bb.mix2 or bb.mix3. 
}
  \item{...}{ other parameters. 
}

}
\author{ Wei Sun weisun2013@gmail.com }
\keyword{ methods }
