\name{QUKtdayX}
\alias{QUKtdayX}
\docType{data}
\title{
QUKtdayX data set
}
\description{
A data frame of 19004 records covering 227 days.
The time stamp ranges from 1 to 96 where each time stamp is a 15-minute interval.

}
\usage{data("QUKtdayX")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{q}{traffic flow, vehicles per hour}
    \item{u}{traffic speed, miles per hour}
    \item{k}{traffic density, vehicles per mile}
    \item{day}{day stamp}
    \item{t}{time-of-day stamp}
    \item{X}{binomial variable for free-flow and congested states}
  }
}
\keyword{datasets}
