% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrtrials.R
\name{rrtrials}
\alias{rrtrials}
\title{\code{rrtrials} produces \code{t-x} trajectories for lead and following vehicles on a ring road}
\usage{
rrtrials(umn, usd, k0, T, dt, leff, nveh, xlim, ylim)
}
\arguments{
\item{umn}{lead vehicle speed (mph), a number}

\item{usd}{standard deviation of lead vehicle speed (mph), a number}

\item{k0}{traffic density, a number}

\item{T}{upper time range, a number}

\item{dt}{time-step, a number}

\item{leff}{effective vehicle length, a number}

\item{nveh}{number of following vehicles, a number}

\item{xlim}{plot range along the \code{t} axis, a vector}

\item{ylim}{plot range along the \code{x} axis, a vector}
}
\value{
The \code{rrtrials} function returns a x-tractories or \code{t-x} plots
for the lead and following vehicles where crossing trajectories are disallowed.
The following vehicle driver maintains a safe headway while attempting to minimize
travel time. In other words, the following vehicle's movement is constrained. Output
from The \code{rrtrials} is an output matrix called \code{LF}.
}
\description{
\code{rrtrials} produces \code{t-x} trajectories for lead and following vehicles on a ring road
}
\examples{
rrtrials(10, 3, 55, 40, 5, 14, 10, c(0, 40), c(-1000, 600))
rrtrials(2, 1, 55, 240, 5, 14, 20, c(0, 240), c(-2000, 600))
}
