% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safecrossover.2R.R
\name{safecrossover2}
\alias{safecrossover2}
\title{The function \code{safecrossover2}  driver conflict of vehicles changing lanes.}
\usage{
safecrossover2(umn, leff)
}
\arguments{
\item{umn}{average speed in mph, a number}

\item{leff}{effective vehicle length in feet, a number}
}
\description{
The function \code{safecrossover2}  driver conflict of vehicles changing lanes.
}
\examples{
safecrossover2(67, 14)
}
