% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbs.simulate.R
\name{sbs.simulate}
\alias{sbs.simulate}
\title{\code{sbs.simulate} uses the law of large numbers to establish confidence intervals traffic performance measures.}
\usage{
sbs.simulate(nveh1,nveh2,umn,usd,xstart1,delt,tstart,tend,xfunnel,leff,size,kfactor)
}
\arguments{
\item{nveh1}{number of vehicles entering the bottleneck from lane 1, a number}

\item{nveh2}{number of vehicles entering the bottleneck from lane 2, a number}

\item{umn}{start speed (mph) for vehicle in lane 1, a number}

\item{usd}{speed standard deviation, a number}

\item{xstart1}{start location of the first vehicle in lane 1, (feet), a number}

\item{delt}{time-step, a number}

\item{tstart}{vehicle crossovers are are permitted below this time, a number}

\item{tend}{upper time range of simulation, a number}

\item{xfunnel}{location where the lane drop is located, a number}

\item{leff}{vehicle length in feet, a number}

\item{size}{sample size, a number}

\item{kfactor}{density at time \code{t} = 0, a number}
}
\value{
The \code{sbs.simulate} returns a list of output.
}
\description{
\code{sbs.simulate} uses the law of large numbers to establish confidence intervals traffic performance measures.
}
