% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{\code{simulate} uses the law of large numbers to establish confidence intervals traffic performance measures.}
\usage{
simulate(nveh1,nveh2,umn,usd,xstart1,xstart2,step,tstart,tend,xfunnel,leff,type,size,browse)
}
\arguments{
\item{nveh1}{number of vehicles entering the bottleneck from lane 1, a number}

\item{nveh2}{number of vehicles entering the bottleneck from lane 2, a number}

\item{umn}{start speed (mph) for vehicle in lane 1, a number}

\item{usd}{speed standard deviation, a number}

\item{xstart1}{start location of the first vehicle in lane 1, (feet), a number}

\item{xstart2}{start location of the first vehicle in lane 2, (feet), a number}

\item{step}{time step, a number}

\item{tstart}{vehicle crossovers are are permitted below this time, a number}

\item{tend}{upper time range of simulation, a number}

\item{xfunnel}{location where the lane drop is located, a number}

\item{leff}{vehicle length in feet, a number}

\item{type}{TRUE to create plots or FALSE otherwise, a logical}

\item{size}{sample size, a number}

\item{browse}{to inspect \code{fixviolation} to inspect plot or FALSE otherwise}
}
\value{
The \code{simulate} runs \code{brktrials4wrapper} \code{size} times.
}
\description{
\code{simulate} uses the law of large numbers to establish confidence intervals traffic performance measures.
}
