% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\docType{data}
\name{size}
\alias{size}
\title{\code{size} sample size.}
\format{\code{size} sample size.
\describe{
\item{size}{a number}
}}
\usage{
size
}
\description{
\code{size} sample size.
}
\keyword{datasets}
