\name{speedts}
\alias{speedts}
\docType{data}
\title{
speedts
}
\description{
\code{speedts} is a subset of \code{QUKtdayX}
}
\usage{data("speedts")}
\format{
  A data frame with 8697 observations on the following 3 variables.
  \describe{
    \item{\code{t}}{a numeric vector}
    \item{\code{u}}{a numeric vector}
    \item{\code{day}}{a numeric vector}
  }
}
\keyword{datasets}
