% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step}
\alias{step}
\title{\code{step} is a time step in a simulation.}
\format{\code{step} is a time step in a simulation, seconds.
\describe{
\item{step}{a number}
}}
\usage{
step(object, scope, scale = 0, direction = c("both", "backward",
  "forward"), trace = 1, keep = NULL, steps = 1000, k = 2, ...)
}
\description{
\code{step} is a time step in a simulation.
}
