\name{t1}
\alias{t1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The following vehicle initiates deceleration.
}
\description{
\code{t1} is time the following vehicle realizes the lead vehicle is decelerating.
}
\value{
\item{t1}{time, measured in seconds.}
}
\author{
Paul J Ossenbruggen
}
