\name{t2}
\alias{t2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The following vehicle sightline intersects the lead vehicle trajectory.
}
\description{
\code{t2} is time the following vehicle realizes the lead vehicle is decelerating.
}

\value{
\item{t2}{time, measured in seconds.}
}
\author{
Paul J Ossenbruggen
}
