\name{t3}
\alias{t3}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
\code{t3} is time the lead vehicle completes deceleration.
}
\description{
The lead vehicle is traveling at \code{u3} = \code{u0}, the breakdown speed at time \code{t3}.
}

\value{
\item{t3}{time, measured in seconds.}
}
\author{
Paul J Ossenbruggen
}
