% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tend.R
\docType{data}
\name{tend}
\alias{tend}
\title{\code{tend} the time \code{t} a simulation ends.}
\format{\code{tend} in seconds.
\describe{
\item{tend}{a number}
}}
\usage{
tend
}
\description{
\code{tend} the time \code{t} a simulation ends.
}
\keyword{datasets}
