\name{theta}
\alias{theta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Exponential deceleration rate \code{theta} of the lead vehicle.
}
\description{
\code{theta} is the decelerattion rate from speed \code{upre} to \code{u0} over time \code{t0} to \code{t3}.
}

\value{
\item{theta}{lead vehicle deceleration rate, measured in feet per second squared.}
}
\author{
Paul J Ossenbruggen
}
