% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstart.R
\docType{data}
\name{tstart}
\alias{tstart}
\title{\code{tstart} \code{t} = 0, a number}
\format{\code{tstart},  generally assigned zero, \code{t} = 0.
\describe{
\item{tstart}{a number}
}}
\usage{
tstart
}
\description{
\code{tstart} \code{t} = 0, a number
}
\keyword{datasets}
