\name{u_sd}
\alias{u_sd}
\docType{data}
\title{
A data set of I-93 speed means (mph) and standard deviation by bin and traffic density (vpm).
}
\description{
A data frame of \code{k}, \code{u} mean and standard deviation data.
}
\usage{data("u_sd")}
\format{
  A data frame with 24 observations on the following 4 variables.
  \describe{
    \item{\code{bin}}{a numeric vector}
    \item{\code{k}}{a numeric vector}
    \item{\code{u.mean}}{a numeric vector}
    \item{\code{u.sd}}{a numeric vector}
  }
}
