\name{uf0}
\alias{uf0}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The following \code{f} vehicle speed \code{uf0} at time \code{t0}.
}
\description{
The following vehicle speed at \code{t0} is assumed to be independent of the lead vehicle speed \code{ul0}.
}
\value{
\item{uf0}{speed, measured in feet per second.}
}
\author{
Paul J Ossenbruggen
}
