\name{ul1}
\alias{ul1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The following \code{l} vehicle speed \code{ul1} at time \code{t1}.
}
\description{
The lead vehicle speed at \code{t1} is estimated with the exponential speed function \code{gbm(upre, theta, t1)}.
}
\value{
 \item{ul1}{speed, measured in feet per second.}
}
\author{
Paul J Ossenbruggen
}
