\name{ul2}
\alias{ul2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The following \code{l} vehicle speed \code{ul2} at time \code{t2}.
}
\description{
The lead vehicle speed at \code{t2} is estimated with the exponential speed function \code{gbm(upre, theta, t2)}.
}
\value{
\item{ul2}{speed, measured in feet per second.}
}
\author{
Paul J Ossenbruggen
}
