\name{ul3}
\alias{ul3}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The following \code{l} vehicle speed \code{ul3} at time \code{t3}.
}
\description{
The lead vehicle speed at \code{t2} is estimated with the exponential speed function \code{gbm(upre, theta, t3)}.
}

\value{
 \item{ul3}{speed, measured in feet per second.}
}
\author{
Paul J Ossenbruggen
}
