% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umn.R
\docType{data}
\name{umn}
\alias{umn}
\title{\code{umn} mean speed of vehicles at time \code{t} = 0.}
\format{\code{umn} speed in miles per hour (mph).
\describe{
\item{umn}{a number}
}}
\usage{
umn
}
\description{
\code{umn} mean speed of vehicles at time \code{t} = 0.
}
\keyword{datasets}
