% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usd.R
\docType{data}
\name{usd}
\alias{usd}
\title{\code{usd}  standard deviation speed of vehicles at time \code{t} = 0.}
\format{\code{usd} given in miles per hour (mph)
\describe{
\item{usd}{a number}
}}
\usage{
usd
}
\description{
\code{usd}  standard deviation speed of vehicles at time \code{t} = 0.
}
\keyword{datasets}
