% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdBBridge.R
\name{usdBBridge}
\alias{usdBBridge}
\title{The \code{usdBBridge} function from Iacus.}
\usage{
usdBBridge(umn, usd, N, T, lambda)
}
\arguments{
\item{umn}{mean speed (mph), a number}

\item{usd}{standard deviation of \code{umn}, a number}

\item{N}{number of time-steps, a number}

\item{T}{upper time range in minutes, a number}

\item{lambda}{calibration constant, a number}
}
\value{
The \code{usdBBridge} function returns information for a single vehicle
used to analyze the effects of traffic noise on car-following.
The model, a stochastic traffic noise model, is assumed to be a Brownian bridge model.
Distance \code{x} is estimated with the forward Euler method.
}
\description{
The \code{usdBBridge} function from Iacus.
}
\examples{
usdBBridge(41, 11, 900, 15, 1.25)
usdBBridge(18.8, 3.9, 900, 15, 0.5)
}
