% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violations.R
\name{violations}
\alias{violations}
\title{\code{violations} produces of safe and unsafe headways for a pair of lead and following vehicles.}
\usage{
violations(df1, df2, leff)
}
\arguments{
\item{df1}{leading vehicle, a matrix}

\item{df2}{following vehicle, a matrix}

\item{leff}{vehicle length, a number}
}
\value{
\code{violations} is used to determine a following vehicle is violating the safe headway rule.
\code{violations} returns a logical, TRUE when a violation occurs and FALSE otherwise.
}
\description{
\code{violations} produces of safe and unsafe headways for a pair of lead and following vehicles.
}
