% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xabparam.R
\name{xabparam}
\alias{xabparam}
\title{Estimate the parameters \code{a} and \code{b} of a second-order velocity model.}
\usage{
xabparam(tstart, tend, ustart, uend, xstart, xend)
}
\arguments{
\item{tstart}{time, a number}

\item{tend}{time, a number}

\item{ustart}{speed, a number}

\item{uend}{speed, a number}

\item{xstart}{location, a number}

\item{xend}{location, a number}
}
\description{
Estimate the parameters \code{a} and \code{b} of a second-order velocity model.
}
