\name{xf1}
\alias{xf1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The lead \code{f} vehicle location \code{xf1} at time \code{t1}.
}
\description{
The following vehicle location at \code{t1} is assumed to be dependent space headway \code{h1}.
}
\value{
\item{xf1}{distance, measured in feet.}
}

\author{
Paul J Ossenbruggen
}

