\name{xf2}
\alias{xf2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The following \code{f} vehicle location \code{xf2} at time \code{t2}.
}
\description{
The following vehicle location at \code{t2} is assumed to be dependent space headway \code{h2}.
}

\value{
' \item{xf2}{distance, measured in feet.}
}

\author{
Paul J Ossenbruggen
}
