\name{xf3}
\alias{xf3}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The lead \code{f} vehicle location \code{xf3} at time \code{t3}.
}
\description{
The following vehicle location at \code{t3} is assumed to be dependent space headway \code{h3}.
}
\value{
\item{xf3}{distance, measured in feet.}
}
\author{
Paul J Ossenbruggen
}

