% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfunnel.R
\docType{data}
\name{xfunnel}
\alias{xfunnel}
\title{\code{xfunnel} upstream location of \code{xe}.}
\format{\code{xfunnel} given in feet
\describe{
\item{xfunnel}{a number}
}}
\usage{
xfunnel
}
\description{
\code{xfunnel} upstream location of \code{xe}.
}
\keyword{datasets}
