% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl0.R
\docType{data}
\name{xl0}
\alias{xl0}
\title{The lead \code{l} vehicle location \code{xl0} at time \code{t0}.}
\format{The lead vehicle location at \code{t0} is assumed to be located upstream of the bottleneck so at time \code{t3} the vehicle reaches \code{x3} = 0.
\describe{
\item{xl0}{distance, measured in feet.}
}}
\usage{
xl0
}
\description{
The lead \code{l} vehicle location \code{xl0} at time \code{t0}.
}
\keyword{datasets}
