% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl2.R
\docType{data}
\name{xl2}
\alias{xl2}
\title{The lead \code{l} vehicle location \code{xl2} at time \code{t2}.}
\format{The lead vehicle location at \code{t2} is estimated with the exponential distance function \code{gbmx(upre, theta, t2)}.
\describe{
\item{xl2}{distance, measured in feet.}
}}
\usage{
xl2
}
\description{
The lead \code{l} vehicle location \code{xl2} at time \code{t2}.
}
\keyword{datasets}
