% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl3.R
\docType{data}
\name{xl3}
\alias{xl3}
\title{The lead \code{l} vehicle location \code{xl3} at time \code{t3}.}
\format{The lead vehicle location at \code{t3} is estimated with the exponential distance function \code{gbmx(upre, theta, t3)}.
\describe{
\item{xl3}{distance, measured in feet.}
}}
\usage{
xl3
}
\description{
The lead \code{l} vehicle location \code{xl3} at time \code{t3}.
}
\keyword{datasets}
