% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xstart1.R
\docType{data}
\name{xstart1}
\alias{xstart1}
\title{\code{xstart1} location of vehicle at \code{t} = tstart}
\format{\code{xstart1} given in feet
\describe{
\item{xstart1}{a number}
}}
\usage{
xstart1
}
\description{
\code{xstart1} location of vehicle at \code{t} = tstart
}
\keyword{datasets}
