% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipper2setup.R
\name{zipper2setup}
\alias{zipper2setup}
\title{\code{zipper2setup} produces a \code{T.} matrix for \code{nveh} vehicles.}
\usage{
zipper2setup(nveh, umn, usd, tstart, tend, xstart, step, leff, kfactor)
}
\arguments{
\item{nveh}{number of vehicles entering the bottleneck, a number}

\item{umn}{start speed (mph) for vehicle in lane 1, a number}

\item{usd}{speed volatility for \code{umn}, a number#' @param tstart start time, (seconds), a number}

\item{tstart}{start time, (seconds), a number}

\item{tend}{end time, (seconds), a number}

\item{xstart}{start location, (feet), a number}

\item{step}{size in seconds, a number}

\item{leff}{vehicle length in feet, a number}

\item{kfactor}{density at time \code{t} = 0, a number}
}
\value{
\code{T.} returns information to plot desire-line trajectories.
}
\description{
\code{zipper2setup} produces a \code{T.} matrix for \code{nveh} vehicles.
}
\examples{
zipper2setup(5, 53.1, 0, 0, 40, -700, 0.125, 14, 1.333)
}
