% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zippermerge.R
\name{zippermerge}
\alias{zippermerge}
\title{\code{zippermerge} produces \code{t-x} trajectory for vehicles traveling on parallel lanes that merge at a bottleneck.}
\usage{
zippermerge(nveh, tstart, tend, xstart, u, leff, xfunnel, step, type, kfactor)
}
\arguments{
\item{nveh}{number of vehicles, a number}

\item{tstart}{start time, (seconds), a number}

\item{tend}{end time, (seconds), a number}

\item{xstart}{vehicle location at time \code{tstart}, a number}

\item{u}{vehicle speed (mph), a number}

\item{leff}{vehicle length in feet, a number}

\item{xfunnel}{location where the lane drop is located, a number}

\item{step}{size in seconds, a number}

\item{type}{TRUE to create plots or FALSE otherwise, a logical}

\item{kfactor}{density at time \code{t} = 0, a number}
}
\value{
The analyses the zone for a safe headway violation. \code{X} = 0 if safe, else non-zero.
}
\description{
\code{zippermerge} produces \code{t-x} trajectory for vehicles traveling on parallel lanes that merge at a bottleneck.
}
