% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoneviolation.R
\name{zoneviolation}
\alias{zoneviolation}
\title{\code{zoneviolation} produces \code{t-x} trajectory for vehicles traveling on parallel lanes that merge at a bottleneck.}
\usage{
zoneviolation(veh, zone, df1df2, dfcrit, delt, tend.0, leff)
}
\arguments{
\item{veh}{identifier, a number}

\item{zone, }{a number}

\item{df1df2}{a data set, a matrix}

\item{dfcrit}{a table of vehicle information, a matrix}

\item{delt}{time-step, a number}

\item{tend.0}{end time for over the long time range, a number}

\item{leff}{vehicle length in feet, a number}
}
\value{
The analyses the zone for a safe headway violation. \code{X} = 0 if safe, else non-zero.
}
\description{
\code{zoneviolation} produces \code{t-x} trajectory for vehicles traveling on parallel lanes that merge at a bottleneck.
}
