% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{\code{anova} returns \code{ANOVA} results, boxplot and \code{Q-Q} plot.}
\usage{
anova(zip, sbs, colnum, figtitle)
}
\arguments{
\item{zip}{output from \code{zipper.simulate}, a data frame}

\item{sbs}{output from \code{sbs.simulate}, a data.frame}

\item{colnum}{column number of interest, a number}

\item{figtitle, }{plot title, character string}
}
\value{
The \code{anova} runs \code{brktrials4wrapper} \code{size} times.
}
\description{
\code{anova} returns \code{ANOVA} results, boxplot and \code{Q-Q} plot.
}
