% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic}
\alias{basic}
\title{Basic interpreter}
\usage{
basic(
  text = NULL,
  file = NULL,
  home_dir = getwd(),
  debug = rly::NullLogger$new()
)
}
\arguments{
\item{text}{Character string containing complete BASIC program (default NULL)}

\item{file}{Name of basic file (default NULL)}

\item{home_dir}{home directory for basic files (default is current working directory)}

\item{debug}{rly debugger}
}
\description{
Basic interpreter
}
\examples{
b <- basic(file='hello.bas', home_dir = system.file('scripts', package = 'basic'))
b$run()
}
