% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmfree2.R
\name{bmfree2}
\alias{bmfree2}
\title{\code{bmfree2} returns a \code{T.} matrix of time \code{t}, speed \code{u} (fps)
 and location \code{x} (feet) using a Brownian Motion Bridge model for one vehicle..}
\usage{
bmfree2(umn, usd, tstart, tend, xstart, step, type)
}
\arguments{
\item{umn}{mean speed (mph), a number}

\item{usd}{standard deviation of \code{usd} (mph), a number}

\item{tstart}{simulation start time, a number}

\item{tend}{simulation end time, a number}

\item{xstart}{location (feet), a number}

\item{step}{time-step size, a number}

\item{browse}{logical for plotting where TRUE creates a plot}
}
\description{
\code{bmfree2} returns a \code{T.} matrix of time \code{t}, speed \code{u} (fps)
 and location \code{x} (feet) using a Brownian Motion Bridge model for one vehicle..
}
