% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brktrials3.R
\name{brktrials3}
\alias{brktrials3}
\title{\code{brktrials3} produces \code{T1} and \code{T2} matrices for lead and following vehicles, respectively}
\usage{
brktrials3(nveh1,nveh2,umn,tstart,tend,xstart1,xstart2,step,browse,leff,xfunnel,usd,kfactor)
}
\arguments{
\item{nveh1}{number of vehicles entering the bottleneck from lane 1, a number}

\item{nveh2}{number of vehicles entering the bottleneck from lane 2, a number}

\item{umn}{start speed (mph) for vehicle in lane 1, a number}

\item{tstart}{start time, (seconds), a number}

\item{tend}{end time, (seconds), a number}

\item{xstart1}{start location of the first vehicle in lane 1, (feet), a number}

\item{xstart2}{start location of the first vehicle in lane 2, (feet), a number}

\item{step}{size in seconds, a number}

\item{browse}{TRUE to create plots or FALSE otherwise, a logical}

\item{leff}{vehicle length in feet, a number}

\item{xfunnel}{upstream location of bottleneck taper, a number}

\item{usd}{speed standard deviation, a number}

\item{kfactor}{density at time \code{t} = 0, a number}
}
\value{
\code{T1} and \code{T2} are returned as a list of two matrices.
}
\description{
\code{brktrials3} produces \code{T1} and \code{T2} matrices for lead and following vehicles, respectively
}
\examples{
brktrials3(5, 5, umn,  0, 40, -700, -700, 0.25, TRUE,  14, -500, 0)
}
