% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialdesign.R
\name{factorialdesign}
\alias{factorialdesign}
\title{\code{factorialdesign} uses the law of large numbers to establish confidence intervals traffic performance measures.}
\usage{
factorialdesign(zipper.output2.5, sbs.output2.5, zipper.output5, sbs.output5)
}
\arguments{
\item{zipper.output2.5}{output from \code{zipper.simulate}, a data frame}

\item{sbs.output2.5}{output from \code{sbs.simulate}, a data.frame}

\item{zipper.output5}{output from \code{zipper.simulate}, a data frame}

\item{sbs.output5}{output from \code{sbs.simulate}, a data.frame}
}
\value{
The \code{factorialdesign} runs \code{brktrials4wrapper} \code{size} times.
}
\description{
\code{factorialdesign} uses the law of large numbers to establish confidence intervals traffic performance measures.
}
