% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{flow}
\alias{flow}
\title{\code{flow} is an algorithm for estimating the flow and density}
\usage{
flow(df1df2, tstart, tend, step, xfunnel, type)
}
\arguments{
\item{df1df2}{vehicle information from \code{brktrials4wrapper}, a matrix}

\item{tstart}{start time, (seconds), a number}

\item{tend}{end time, (seconds), a number}

\item{step}{size in seconds, a number}

\item{xfunnel}{upstream location of bottleneck taper, a number}

\item{type}{TRUE to create plots or FALSE otherwise, a logical}
}
\value{
\code{flow} returns data.frame of density and flow estimates at two bottleneck locations.
}
\description{
\code{flow} is an algorithm for estimating the flow and density
}
