% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsafe.R
\name{hsafe}
\alias{hsafe}
\title{Safe headway \code{h} between lead and following vehicles. Speed \code{u} of lead vehicle is measured in mph. \code{leff} = 14 feet.}
\usage{
hsafe(u, leff)
}
\arguments{
\item{u}{speed in units of fps, a number}

\item{leff}{effective vehicle length in feet, a number}
}
\description{
Safe headway \code{h} between lead and following vehicles. Speed \code{u} of lead vehicle is measured in mph. \code{leff} = 14 feet.
}
