% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leff.R
\docType{data}
\name{leff}
\alias{leff}
\title{\code{leff} vehicle length in feet.}
\format{\code{leff} a number.
\describe{
\item{leff}{a number}
}}
\usage{
leff
}
\description{
\code{leff} vehicle length in feet.
}
\keyword{datasets}
