% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{\code{performance} looks at waiting time and performance measures.}
\usage{
performance(run1,run2,p1,p2)
}
\arguments{
\item{run1, }{a matrix}

\item{run2, }{a matrix}

\item{p1}{proportion free flow, a number}

\item{p2}{proportion free flow, a number}
}
\value{
The \code{performance} returns scatter plots, histograms and statistical summaries for two data sets.
}
\description{
\code{performance} looks at waiting time and performance measures.
}
