% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{\code{performance} looks at waiting time and performance measures.}
\usage{
performance(run1,run2)
}
\arguments{
\item{run1, }{a data frame}

\item{run2, }{a data frame}
}
\value{
The \code{performance} returns scatter plots, histograms and statistical summaries for two data sets.
}
\description{
\code{performance} looks at waiting time and performance measures.
}
