% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedplot.R
\name{speedplot}
\alias{speedplot}
\title{The function \code{speedplot} plots the data from a data frame \code{df}}
\usage{
speedplot()
}
\description{
The function \code{speedplot} plots the data from a data frame \code{df}
}
