\name{xl0}
\alias{xl0}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The lead \code{l} vehicle speed \code{ul0} at time \code{t0}.
}
\description{
The lead vehicle speed at \code{t0} is assumed to be independent of the pre-breakdown speed \code{upre}.
}

\value{
\item{xl0}{location, measured in feet.}
}
\author{
Paul J Ossenbruggen
}
