% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl1.R
\docType{data}
\name{xl1}
\alias{xl1}
\title{The lead \code{l} vehicle location \code{xl1} at time \code{t1}.}
\format{The lead vehicle location at \code{t1} is estimated with the exponential distance function \code{gbmx(upre, theta, t1)}.
\describe{
\item{xl1}{distance, measured in feet.}
}}
\usage{
xl1
}
\description{
The lead \code{l} vehicle location \code{xl1} at time \code{t1}.
}
\keyword{datasets}
